feat(upodder): Adding user confirmation before downloading and force episode downloading
* Adding user confirmation before downloading. 
Before upodder downloads an episode it will ass the user for confirmation. This way the user can avoid downloading an episode that they don't want to download. This can be prevented with the --no-confirm flag, which will revert to the previous behavior. 
* Adding force episode downloading.
Using the --episode  flag the user can specify an episode title to force download. This means that if a user wants to download an individual episode that is already seen, they can do so.